requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
opentelemetry-api==1.36.0
opentelemetry-exporter-prometheus==0.57b0
opentelemetry-instrumentation==0.57b0
opentelemetry-instrumentation-flask==0.57b0
opentelemetry-instrumentation-wsgi==0.57b0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
opentelemetry-util-http==0.57b0
packaging==25.0
prometheus_client==0.22.1
typing_extensions==4.14.1
Werkzeug==3.1.3
wrapt==1.17.3
zipp==3.23.0
